Rename file according to class name
