+++ src/radarewebserver.cpp
@@ -1,4 +1,4 @@
-#include "webserverthread.h"
+#include "radarewebserver.h"
 #include "qrcore.h"
 #include <cassert>
